Updates on jes and BTag treatment

+++ AnalysisStep/test/Ntuplizers/HZZ4lNtupleMaker.cc
@@ -306,7 +306,7 @@ namespace {
 
   std::vector<float> JetPUValue;
   std::vector<short> JetPUID;
-  std::vector<short> JetPUID_score;
+  std::vector<float> JetPUID_score;
     
   std::vector<short> JetID;
